sdk/python/requirements.txt
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
fastapi>=0.100.0
